refactor: extract components from PostPortableText
