+++ src/components/GenericPagesComponents/PostBlockCallout/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./PostBlockCallout";
