+++ src/components/GenericPagesComponents/PostCta/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./PostCta";
