+++ src/components/GenericPagesComponents/PostFootnotes/index.tsx
@@ -1,0 +1,6 @@
+export {
+  extractFootnotes,
+  PostFootnoteAnnotation,
+  PostFootnotesSection,
+} from "./PostFootnotes";
+export type { Footnote } from "./PostFootnotes";
