+++ src/components/GenericPagesComponents/PostVideo/PostVideo.tsx
@@ -1,7 +1,6 @@
 import { PortableTextComponentProps } from "@portabletext/react";
 
-import { Video } from "../../../common-lib/cms-types";
-
+import { Video } from "@/common-lib/cms-types";
 import CMSVideo from "@/components/SharedComponents/CMSVideo";
 import Box from "@/components/SharedComponents/Box";
 import Flex from "@/components/SharedComponents/Flex";
