+++ src/components/GenericPagesComponents/PostVideo/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./PostVideo";
